fixed line_item decimal/integer bug

+++ db/migrate/20120519133247_create_line_items.rb
@@ -2,7 +2,7 @@ class CreateLineItems < ActiveRecord::Migration
   def change
     create_table :line_items do |t|
       t.decimal :unit_price
-      t.decimal :product_id
+      t.integer :product_id
       t.integer :cart_id
       t.integer :quantity
